refactor admin page to include content block
